refactor complex function
