+++ src/components/activity-groups-list/activity-group-item/activity-group-type/activity-group-type.tsx
@@ -1,16 +1,14 @@
 import React, { FC, useMemo } from 'react';
 import { Text, View } from 'react-native';
 
-import { ActivityTypeEnum } from '../../../../enums/activity-type.enum';
-import { ActivityGroup, emptyActivity } from '../../../../interfaces/activity.interface';
+import { ActivityGroup } from '../../../../interfaces/activity.interface';
 import { useSelectedAccountSelector } from '../../../../store/wallet/wallet-selectors';
 import { formatSize } from '../../../../styles/format-size';
 import { useColors } from '../../../../styles/use-colors';
-import { isString } from '../../../../utils/is-string';
 import { Divider } from '../../../divider/divider';
 import { Icon } from '../../../icon/icon';
-import { IconNameEnum } from '../../../icon/icon-name.enum';
 import { useActivityGroupTypeStyles } from './activity-group-type.styles';
+import { handleChooseIcon, IconGroupType } from './activity-group-type.util';
 
 interface Props {
   group: ActivityGroup;
@@ -22,40 +20,10 @@ export const ActivityGroupType: FC<Props> = ({ group }) => {
 
   const publicKeyHash = useSelectedAccountSelector().publicKeyHash;
 
-  const [iconName, iconColor, typeText] = useMemo<[IconNameEnum, string, string]>(() => {
-    if (group.length > 1) {
-      return [IconNameEnum.Clipboard, colors.gray1, 'Interaction'];
-    }
-
-    const firstActivity = group[0] ?? emptyActivity;
-
-    if (firstActivity.type === ActivityTypeEnum.Transaction) {
-      if (firstActivity.source.address === publicKeyHash) {
-        return [
-          IconNameEnum.ArrowUp,
-          colors.destructive,
-          isString(firstActivity.entrypoint)
-            ? `Called ${firstActivity.entrypoint}`
-            : firstActivity.destination.alias ?? 'Sent'
-        ];
-      } else {
-        return [IconNameEnum.ArrowDown, colors.adding, firstActivity.source.alias ?? 'Received'];
-      }
-    } else {
-      if (firstActivity.type === ActivityTypeEnum.Delegation) {
-        const alias = firstActivity.destination.alias;
-        const postfix = isString(alias) ? ` to ${alias}` : '';
-
-        return [
-          IconNameEnum.Deal,
-          colors.gray1,
-          isString(firstActivity.destination.address) ? 'Delegated' + postfix : 'Undelegated'
-        ];
-      } else {
-        return [IconNameEnum.Clipboard, colors.gray1, 'Undelegated'];
-      }
-    }
-  }, [group, publicKeyHash, colors]);
+  const [iconName, iconColor, typeText] = useMemo<IconGroupType>(
+    () => handleChooseIcon(group, colors, publicKeyHash),
+    [group, publicKeyHash, colors]
+  );
 
   return (
     <View style={styles.container}>
